Moves ring/circular buffer implementation into its own header in the data structure module
