+++ .gitignore
@@ -63,3 +63,5 @@ compile_commands.json
 #special test files
 *test.txt
 *shader.hlsl
+tags
+imgui.ini
