+++ src/Core/DataStructures/RingeBuffer.cpp
@@ -1,0 +1,8 @@
+#include "RingBuffer.h"
+
+namespace Kraid
+{
+
+
+
+}
